UBUNTU: [Config] move ECRYPT_FS back to =y for all architectures
ecryptfs is non-autoloadable and so we need ECRYPT_FS=y.  ECRYPT_FS has
recently gained support for trusted keys.  This adds a dependancy on
ENCRYPTED_KEYS which in turn has a dependancy on TRUSTED_KEYS which in
turn is dependant on CONFIG_TCG_TPM.  In order to have ECRYPT_FS built
in we have to build all of these others too.

BugLink: http://bugs.launchpad.net/bugs/827197
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>