UBUNTU: [Config] clean up the human consumable package descriptions
Each flavour has a human readable description in its flavour vars.
This is intended to describe the system type in words, for example as
here for powerpc:

    desc="32-bit PowerPC"
    desc="32-bit PowerPC SMP"
    desc="64-bit PowerPC SMP"

On x86 we commonly share flavour names and there this is currently
rather unhelp:

    desc="x86/x86_64"

Additionally for these shared flavours is it impossible to have an accurate
description as it must include all possible architectures, even though
the resultant description is on an architecture specific binary package.

Introduce a human_arch configuration at the archtecture level, this can
then be used as a component in the flavour description.  Use this to
add descriptions for the x86 architectures:

    human_arch	= 64 bit x86
    human_arch	= 32 bit x86

Allowing descriptions as below to be specified:

    desc="=HUMAN= SMP"

Resulting in full descriptions as below:

    Linux kernel image for version 3.2.0 on 64 bit x86 SMP

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>